old-tests/submission/Infraware/Forms/contents/Forms/Not Use/input_multiple.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_multiple</h3>
	</p>
	
	<li>You can select multiple files.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><input type='file' id='input_file' onchange='change()' multiple></p>
	</form>

	<script>
		var t1 = async_test('input multiple test');
		
		function change() {
			var input_file = document.getElementById("input_file");
			var multiple = input_file.multiple;

			if (typeof(multiple) == "boolean") {
				t1.step(function() {
					assert_equals(multiple, true, "multiple attribute is not correct.");
				});
			} else {
				t1.step(function() {
					assert_unreached("multiple attribute is not exist.");
				});
			}
			t1.done();
		}

	</script>

 </body>
</html>
